refactored styled components to separate files
