+++ src/components/ListingCard/ListingCard.tsx
@@ -1,38 +1,8 @@
 import React from 'react';
-import styled from 'styled-components';
 import { UUID } from 'crypto';
-import { CaseListing } from '../types/schemaTypes';
-import timestampStringToDate from '../utils/helpers';
-import CardTag from './CardTag';
-
-// styled components
-const CardBody = styled.div<{ $selected?: boolean }>`
-  display: flex;
-  flex-direction: column;
-  border: 1px solid lightgray;
-  padding: 1rem;
-  border-radius: 10px;
-  transition: 150ms;
-  cursor: pointer;
-  gap: 1rem;
-
-  ${({ $selected }) => $selected && `border-color: #097A62`};
-
-  &:hover {
-    box-shadow: 0 5px 5px rgba(0, 0, 0, 0.1);
-    transform: translateY(-2px);
-  }
-`;
-
-const CardTitle = styled.h2`
-  font-size: 1.5rem;
-  margin: 0;
-`;
-
-const TagRow = styled.div`
-  display: flex;
-  gap: 1rem;
-`;
+import { CaseListing } from '../../types/schemaTypes';
+import timestampStringToDate from '../../utils/helpers';
+import { CardBody, CardTitle, TagRow, CardTag } from './styles';
 
 export default function ListingCard({
   caseData,
@@ -75,9 +45,10 @@ export default function ListingCard({
         {rolesNeeded.map(r => (
           <CardTag
             key={r}
-            text={r}
             color={r === 'Interpreter' ? interpreterColor : attorneyColor}
-          />
+          >
+            {r}
+          </CardTag>
         ))}
       </TagRow>
     </CardBody>
